+++ NEWS.md
@@ -31,6 +31,9 @@ CHANGES
 * The `regex` argument was added to functions that use select-helpers and did
   not already have this argument.
   
+* Select helpers `starts_with()`, `ends_with()`, and  `contains()` now accept
+  several patterns, e.g `starts_with("Sep", "Petal")`.
+  
 * Arguments `select` and `exclude` that are present in most functions have been
   improved to work in loops and in custom functions. For example, the following 
   code now works: